+++ src/esengine/renderer/frame/RenderFrame.hpp
@@ -142,12 +142,10 @@ public:
     void processMasks(ecs::Registry& registry, i32 vpX, i32 vpY, i32 vpW, i32 vpH);
 
     void setEntityClipRect(u32 entity, i32 x, i32 y, i32 w, i32 h);
-    void clearEntityClipRect(u32 entity);
     void clearAllClipRects();
 
     void setEntityStencilMask(u32 entity, i32 refValue);
     void setEntityStencilTest(u32 entity, i32 refValue);
-    void clearEntityStencilMask(u32 entity);
     void clearAllStencilMasks();
 
     void submitTileQuad(